+++ src/Control/Monad/Trans/Indexed/Cont.hs
@@ -1,3 +1,12 @@
+{- |
+Module      :  Control.Monad.Trans.Indexed.Cont
+Copyright   :  (C) 2024 Eitan Chatav
+License     :  BSD 3-Clause License (see the file LICENSE)
+Maintainer  :  Eitan Chatav <eitan.chatav@gmail.com>
+
+The continuation indexed monad transformer.
+-}
+
 module Control.Monad.Trans.Indexed.Cont
   ( ContIx (..)
   , callCCIx