+++ src/Control/Monad/Trans/Indexed/Free.hs
@@ -1,6 +1,15 @@
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -fno-warn-name-shadowing #-}
 
+{- |
+Module      :  Control.Monad.Trans.Indexed.Free
+Copyright   :  (C) 2024 Eitan Chatav
+License     :  BSD 3-Clause License (see the file LICENSE)
+Maintainer  :  Eitan Chatav <eitan.chatav@gmail.com>
+
+The free indexed monad transformer.
+-}
+
 module Control.Monad.Trans.Indexed.Free
   ( IxMonadTransFree (liftFreeIx, hoistFreeIx, foldFreeIx), coerceIxMonadTransFree
   , IxFunctor, IxMap (IxMap), liftFreeIxr, hoistFreeIxr