+++ ChangeLog.md
@@ -1,3 +1,9 @@
 # Changelog for indexed-transformers
 
-## Unreleased changes
+* 0.0.1
+  - indexed monad transformers
+  - free indexed monad transformer
+  - continuation indexed monad transformer
+  - state indexed monad transformer
+  - writer indexed monad transformer
+  - qualified indexed do notation